+++ sequential/fibonacci/main.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 
 
-int fibonacci(int input) {
+long long fibonacci(long long input) {
     if (input < 2 ) return input;
     return fibonacci(input - 1) + fibonacci(input - 2);
 }
@@ -21,7 +21,7 @@ int main(int argc, char *argv[]) {
     auto start = std::chrono::high_resolution_clock::now();
 
     // call fibonacci function
-    int result = fibonacci(fibNumber);
+    long long result = fibonacci((long long) fibNumber);
 
     auto end = std::chrono::high_resolution_clock::now();
     auto duration = std::chrono::duration_cast<std::chrono::microseconds>(end - start);
